requirements.txt
pyyaml
keyboard
pyautogui
requests
pyserial
comtypes
pystray
pillow
https://github.com/kdschlosser/pyWinCoreAudio/tarball/develop
